+++ data/import_visits.sql
@@ -10,7 +10,7 @@ WITH module AS (
     WHERE module_code ILIKE :'moduleCode'
 ),
 sites AS (
-    SELECT id_base_site AS id, base_site_code::int AS site_code
+    SELECT id_base_site AS id, base_site_code AS site_code
     FROM gn_monitoring.t_base_sites
 )
 UPDATE :moduleSchema.:visitsTmpTable AS vt
